fix(frontend): adress yarn build failed problem
